requirements.txt
pyngrok
pyco
PySimpleGUI
PyCryptodome
